Project_2/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
databases
passlib
python-jose
kafka-python
python-jose[cryptography]